+++ arch/arm/mach-omap2/voltage.h
@@ -88,7 +88,11 @@ struct voltagedomain {
 
 	int (*scale) (struct voltagedomain *voltdm,
 		      unsigned long target_volt);
-	unsigned long (*get_voltage) (struct voltagedomain *voltdm);
+	/*
+	 * FIXME should this be a long to return error codes?
+	 * e.g. if VC cannot support get_voltage, return -EPERM
+	 */
+	u32 (*get_voltage) (struct voltagedomain *voltdm);
 
 	u32 nominal_volt;
 	struct omap_volt_data *volt_data;
